refactor: shorten public type names and tighten result types
Drop the RedosDetector prefix from the public types. Through the crate
path the prefix stuttered: `redos_detector::RedosDetectorTrail` reads worse
than `redos_detector::Trail`. New names: NodeLocation, Node, BackReference,
QuantifierIteration, TrailEntrySide, TrailEntry, Trail, AnalysisLimit (the
analysis-limit enum), Report (the result).

Tighten the result shape:
- Drop the `safe` field. It duplicated `error.is_none()` and could drift
  out of sync. Add `Report::is_safe()` in its place.
- Rename `Trail.trail` to `Trail.entries` so access reads `trail.entries`
  instead of `trail.trail`.

Make `Score` carry a whole count. A score is a path count, so `Finite(u64)`
fits and lets `Score` derive `Eq` and `Hash`. Callers no longer write float
literals like `Score::Finite(1.0)`.

Mark `AnalysisLimit` and `Score` `#[non_exhaustive]` so a future variant is
not a breaking change. Match them with a wildcard arm.

+++ tests/cases_data/mod.rs
@@ -5,7 +5,7 @@ pub use trail_counts::TRAIL_COUNTS;
 pub enum Exp {
     Safe,
     Unsafe,
-    Score(f64),
+    Score(u64),
 }
 
 pub struct Case {
@@ -501,8 +501,8 @@ pub const CASES: &[Case] = &[
     Case { source: "\\s+", flags: "", expected: Exp::Safe, missing_anchor: true, infinite: false },
     Case { source: "(a+)+", flags: "", expected: Exp::Safe, missing_anchor: true, infinite: false },
     Case { source: "(a+)+$", flags: "", expected: Exp::Unsafe, missing_anchor: true, infinite: false },
-    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8})*)+)?$", flags: "", expected: Exp::Score(2.0), missing_anchor: false, infinite: false },
-    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8}){0,5})+)?$", flags: "", expected: Exp::Score(2.0), missing_anchor: false, infinite: false },
-    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8}){0,2})+)?$", flags: "", expected: Exp::Score(2.0), missing_anchor: false, infinite: false },
-    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8}){0,2}){1,5})?$", flags: "", expected: Exp::Score(2.0), missing_anchor: false, infinite: false },
+    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8})*)+)?$", flags: "", expected: Exp::Score(2), missing_anchor: false, infinite: false },
+    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8}){0,5})+)?$", flags: "", expected: Exp::Score(2), missing_anchor: false, infinite: false },
+    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8}){0,2})+)?$", flags: "", expected: Exp::Score(2), missing_anchor: false, infinite: false },
+    Case { source: "^(?:-(?:[a-z][a-z\\d]{4,7}|\\d[a-z\\d]{3,7}))*(?:_[^x](?:_[a-z\\d]{2}(_[a-z\\d]{3,8}){0,2}){1,5})?$", flags: "", expected: Exp::Score(2), missing_anchor: false, infinite: false },
 ];
